pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chakra-research"
version = "0.3.0"
description = "Chakra — Autonomous Research System. A cyclic, domain-agnostic ML research engine."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
  {name = "Harsh Vardhan"},
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
  "pillow>=10.0.0",
  "pyyaml>=6.0.0",
  "scikit-image>=0.22.0",
  "scikit-learn>=1.4.0",
  "torch>=2.2.0",
  "torchvision>=0.17.0",
  "tqdm>=4.66.0",
  "wandb>=0.17.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
autoresearch = "chakra.cli:main"
chakra = "chakra.chakra_cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"chakra.domains.hndsr_vr" = ["domain.yaml"]
"chakra.domains.nlp_lm" = ["domain.yaml"]
"chakra.domains.tabular_cls" = ["domain.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q -p no:cacheprovider"
